+++ src/main.rs
@@ -31,7 +31,6 @@ extern crate lazy_static;
 
 #[cfg(windows)]
 use std::ptr::write_volatile;
-use std::sync::Arc;
 
 use anyhow::Result;
 use clap::Parser;
@@ -103,38 +102,4 @@ pub fn setup_logging(clargs: &Configuration) {
                 .init();
         }
     }
-}
-
-/// Initializes the authentication module and cookie store and builds a Reqwest HTTP client
-fn build_http_client(
-    api: &openapiv3::OpenAPI,
-) -> Result<
-    (
-        authentication::Authentication,
-        Arc<reqwest_cookie_store::CookieStoreMutex>,
-        reqwest::blocking::Client,
-    ),
-    anyhow::Error,
-> {
-    let server_url = reqwest::Url::parse(
-        &api.servers
-            .first()
-            .ok_or(anyhow!("Could not extract server URL from API spec"))?
-            .url,
-    )?;
-
-    // Load auth information from the configuration
-    let mut authentication = authentication::initialize()?;
-    // Make a cookie jar for our client
-    let cookie_store = std::sync::Arc::new(reqwest_cookie_store::CookieStoreMutex::new(
-        authentication.cookie_store(&server_url)?,
-    ));
-    // Construct a client with the authentication and static headers
-    let client_builder =
-        reqwest::blocking::Client::builder().cookie_provider(std::sync::Arc::clone(&cookie_store));
-    let mut default_headers = authentication.generate_headers();
-    default_headers.extend(header::get_default_headers()?);
-    let client = client_builder.default_headers(default_headers).build()?;
-
-    Ok((authentication, cookie_store, client))
 }