+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 - Updated [LibAFL](https://github.com/AFLplusplus/LibAFL) to 0.15.0 in
   [#65](https://github.com/TNO-S3/WuppieFuzz/pull/65) thanks to
   [@tokatoka](https://github.com/tokatoka)
+- Sqlite reporting: save body contents as text (instead of `&[u8]`) in
+  [#76](https://github.com/TNO-S3/WuppieFuzz/pull/76/)
 
 # v1.1.2 (2025-01-16)